Bug Fix
The muppet that is me strikes again. I forgot to let the LRU algo when I was using stuff correctly. As such it was basically FIFO.

+++ src/memoryAlgorithms/LRU.java
@@ -20,7 +20,7 @@ public class LRU implements Algorithm {
 	public void used(int address) {
 		
 		if(used.contains(address)){
-			used.remove(address);
+			used.remove( (Integer) address);
 		}
 		used.add(address);
 	}
